Use duck-typing for BigNumber when instanceof fails
When using this library across JavaScript contexts, an object may
be an instanceof a BigNumber, but return false to
`instanceof BigNumber`. This occurs when the BigNumber was created
in a different JavaScript context.

This can happen in a Node environment if an application has a
dependency for `bignumber.js`. Depending on how the modules are
installed, `json-bigint` can use a different copy than the main
application. Then any BigNumber instances passed into `stringify`
aren't recognized as such and are formatted as strings instead.