Use BigNumber.isBigNumber
Switch to use the new BigNumber.isBigNumber property to determine
if an object is a BigNumber. This was introduced in BigNumber 4.0.0
and is more reliable than duck-typing the `c`, `e`, and `s`
properties. Keep an instanceof so this works with older versions of
the BigNumber library if only used in a single JavaScript context.